1 - Removed Rare Variant rarity, replaced with Challenge Drop tooltip. 2 - Replaced Pharaoh Set in Pyramids with a Ruby Hook via IL Editing. 3 - Gave several split projectiles homing a set time after the primary projectile splits, these projectiles deal no damage for a short time and then begin homing once that time has passed. 4 - Reduced Rocket III and Rocket IV damage by 25%, buffed low damage launchers a bit to compensate. 5 - Reverted the nerf made to Snowman Cannon. 6 - Adjusted Scorpio to be a burst-fire weapon. 7 - Increased homing speed for many weapons but reduced homing range. 8 - Reworked Conference Call to target up to 12 enemies on your screen and fire bullets from the sky towards them. 9 - Crown Jewel and Teardrop Cleaver are now Challenge Drops. 10 - Magic Mirror, Ice Mirror, Recall Potions and Cell Phone have a tooltip to state that they can't be used during boss fights. 11 - Reduced Bohldohr damage and made them spawn more often. 12 - Adjusted Gem Crawler spawn rates. 13 - Earth Elemental drop pool changes and made them non-hostile and non-damageable unless approached. 14 - Providence no longer fires additional stars directly at the player given her guardians are alive and she's not at low HP. 15 - Pillar event spawns are now tiered, weaker enemies spawn while the shield is at high HP and more difficult enemies spawn as the HP lowers. 16 - Demon Altars drop 6 Souls of Night upon being broken. 17 - Added recipes for Pharaoh Set pieces.

+++ Projectiles/Ranged/OnyxSharkBomb.cs
@@ -85,8 +85,8 @@ namespace CalamityMod.Projectiles.Ranged
             for (i = 0; i < 2; i++)
             {
                 offsetAngle = startAngle + deltaAngle * (i + i * i) / 2f + 32f * i;
-                Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, (float)(Math.Sin(offsetAngle) * 5f), (float)(Math.Cos(offsetAngle) * 5f), ModContent.ProjectileType<SandyWaifuShark>(), (int)((double)projectile.damage * 0.4), projectile.knockBack, projectile.owner, 0f, 0f);
-                Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, (float)(-Math.Sin(offsetAngle) * 5f), (float)(-Math.Cos(offsetAngle) * 5f), ModContent.ProjectileType<SandyWaifuShark>(), (int)((double)projectile.damage * 0.4), projectile.knockBack, projectile.owner, 0f, 0f);
+                Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, (float)(Math.Sin(offsetAngle) * 5f), (float)(Math.Cos(offsetAngle) * 5f), ModContent.ProjectileType<SandyWaifuShark>(), (int)(projectile.damage * 0.4), projectile.knockBack, projectile.owner, 0f, 0f);
+                Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, (float)(-Math.Sin(offsetAngle) * 5f), (float)(-Math.Cos(offsetAngle) * 5f), ModContent.ProjectileType<SandyWaifuShark>(), (int)(projectile.damage * 0.4), projectile.knockBack, projectile.owner, 0f, 0f);
             }
         }
     }
